moving translation data tags clean-up functions to data_utils
